+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
- ## [Unreleased]
+## [Unreleased]
 
- ### Added
+### Added
 
+- [TD-5947] Support for DataViews
 - [TD-5495] Foreing keys columns should match original ID columns in all tables
 
 ## [5.14.0] 2023-09-19